Replaced taxon_index_to_html with taxon_index_to_text
