+++ safedata_validator/taxa.py
@@ -40,6 +40,8 @@ from logging import Formatter
 from typing import Optional, Union
 
 import requests  # type: ignore
+from dominate import tags  # TEMP - DELETE LATER
+from dominate.util import raw  # TEMP - DELETE LATER
 from lxml import etree
 from openpyxl import worksheet
 
@@ -2786,9 +2788,9 @@ class Taxa:
         return self.gbif_taxa.taxon_names.intersection(self.ncbi_taxa.taxon_names)
 
 
-# TODO - Decide whether to delete this function
+# TODO - Fix multiple branch error
 def taxon_index_to_text(
-    taxon_index: list, html: bool = False, indent_width: int = 4
+    taxa: list[dict], html: bool = False, indent_width: int = 4
 ) -> str:
     """Render a GBIFTaxa instance as text or html.
 
@@ -2800,16 +2802,19 @@ def taxon_index_to_text(
         taxon_index: The taxon_index property of a GBIFTaxa or NCBITaxa instance.
         html: Render as html or text
         indent_width: The indentation width to use for successive taxonomic ranks.
+
+    Returns:
+        A `dominate.tags.div` object containing an HTML representation of the taxa.
     """
 
     lbr = "<br>" if html else "\n"
 
-    def indent(n, use_html=html):
+    def _indent(n, use_html=html):
 
         ind = "&ensp;" if use_html else " "
         return ind * indent_width * (n - 1)
 
-    def format_name(tx, use_html=html):
+    def _format_name(tx, use_html=html):
 
         # format the canonical name
         if tx[4] in ["genus", "species", "subspecies"]:
@@ -2825,55 +2830,170 @@ def taxon_index_to_text(
     # Container to hold the output
     html_out = StringIO()
 
-    # group by parent taxon id in position 2, substituting 0 for None
-    taxon_index.sort(key=lambda x: x[2] or 0)
-    grouped = {k: list(v) for k, v in groupby(taxon_index, lambda x: x[2])}
+    # group by parent taxon, substituting 0 for None
+    taxa.sort(key=lambda x: x["gbif_parent_id"] or 0)
+    grouped = {k: list(v) for k, v in groupby(taxa, lambda x: x["gbif_parent_id"])}
 
     # start the stack with the kingdoms - these taxa will have None as a parent
-    stack = [{"current": grouped[None][0], "next": grouped[None][1:]}]
+    stack = [({"current": grouped[None][0]}, {"next": grouped[None][1:]})]
 
     while stack:
 
         # Handle the current top of the stack: format the canonical name
-        current = stack[-1]["current"]
-        canon_name = format_name(current)
+        current = stack[-1][0]["current"]
+        canon_name = _format_name(current)
 
         # Look for a non-None entry in next that shares the same worksheet name
-        next_ws_names = [tx[0] for tx in stack[-1]["next"] if tx[0] is not None]
+        next_ws_names = [
+            tx["worksheet_name"]
+            for tx in stack[-1][1]["next"]
+            if tx["worksheet_name"] is not None
+        ]
 
-        if current[0] in next_ws_names:
+        if current["worksheet_name"] in next_ws_names:
             # pop out the matching entry and find which is 'accepted'
-            name_pair = stack[-1]["next"].pop(next_ws_names.index(current[0]))
-            if current[5] == "accepted":
-                as_name = format_name(name_pair)
-                as_status = name_pair[5]
+            name_pair = stack[-1][1]["next"].pop(
+                next_ws_names.index(current["worksheet_name"])
+            )
+            if current["gbif_status"] == "accepted":
+                as_name = _format_name(name_pair)
+                as_status = name_pair["gbif_status"]
             else:
                 as_name = canon_name
-                as_status = current[5]
-                canon_name = format_name(name_pair)
+                as_status = current["gbif_status"]
+                canon_name = _format_name(name_pair)
 
-            txt = f"{indent(len(stack))} {canon_name} (as {as_status}: {as_name}){lbr}"
+            txt = f"{_indent(len(stack))} {canon_name} (as {as_status}: {as_name}){lbr}"
         else:
-            txt = f"{indent(len(stack))} {canon_name}{lbr}"
+            txt = f"{_indent(len(stack))} {canon_name}{lbr}"
 
         html_out.write(txt)
 
         # Is this taxon a parent for other taxa - if so add that taxon to the top of
         # the stack, otherwise start looking for a next taxon to push onto the stack.
         # If there is none at the top, pop and look down.
-        parent_id = current[1]
+        parent_id = current["gbif_taxon_id"]
         if parent_id in grouped:
             stack.append(
-                {"current": grouped[parent_id][0], "next": grouped[parent_id][1:]}
+                ({"current": grouped[parent_id][0]}, {"next": grouped[parent_id][1:]})
             )
         else:
             while stack:
                 push = stack.pop()
-                if push["next"]:
-                    stack.append({"current": push["next"][0], "next": push["next"][1:]})
+                if push[1]["next"]:
+                    stack.append(
+                        ({"current": push[1]["next"][0]}, {"next": push[1]["next"][1:]})
+                    )
                     break
 
     return html_out.getvalue()
+
+
+def ncbi_index_to_html(taxa: list[dict]) -> tags.div:
+    """Generate an HTML formatted taxon list for an NCBI index.
+
+    Takes a taxon index - a list containing taxon dictionaries - and converts into
+    an dominate.tags.div() object containing a simple HTML representation of the
+    taxonomy. The representation uses indentation to show taxonomic depth. This function
+    differs from `taxon_index_to_html` in that it works for NCBI indices rather than
+    GBIF.
+
+    Arguments:
+        taxa: A list of taxon dictionaries containing the taxa for a dataset.
+
+    Returns:
+        A `dominate.tags.div` object containing an HTML representation of the taxa.
+    """
+
+    def _indent(n):
+
+        return raw("&ensp;-&ensp;" * n)
+
+    def _format_name(tx):
+
+        # format the canonical name
+        if tx["ncbi_status"] == "user":
+            if tx["taxon_rank"] in BACKBONE_RANKS_EX:
+                return f"[{tx['taxon_name']}]"
+            else:
+                return f"[{tx['taxon_name']}]  (non-backbone rank: {tx['taxon_rank']})"
+        else:
+            if tx["taxon_rank"] in ["genus", "species", "subspecies"]:
+                return tags.i(tx["taxon_name"])
+            elif tx["taxon_rank"] not in BACKBONE_RANKS_EX:
+                return f"{tx['taxon_name']} (non-backbone rank: {tx['taxon_rank']})"
+            else:
+                return tx["taxon_name"]
+
+    # Container to hold the output
+    html = tags.div()
+
+    # group by parent taxon, substituting 0 for None
+    taxa.sort(key=lambda x: x["ncbi_parent_id"] or 0)
+    grouped = {k: list(v) for k, v in groupby(taxa, lambda x: x["ncbi_parent_id"])}
+
+    # start the stack with the superkingdoms - these taxa will have None as a parent
+    stack = [({"current": grouped[None][0]}, {"next": grouped[None][1:]})]
+
+    while stack:
+
+        # Handle the current top of the stack: format the canonical name
+        current = stack[-1][0]["current"]
+        canon_name = _format_name(current)
+
+        # Look for a non-None entry in next that shares the same worksheet name
+        next_ws_names = [
+            tx["worksheet_name"]
+            for tx in stack[-1][1]["next"]
+            if tx["worksheet_name"] is not None
+        ]
+
+        if current["worksheet_name"] in next_ws_names:
+            # pop out the matching entry and find which is 'accepted'
+            name_pair = stack[-1][1]["next"].pop(
+                next_ws_names.index(current["worksheet_name"])
+            )
+            if current["ncbi_status"] == "accepted":
+                as_name = _format_name(name_pair)
+                as_status = name_pair["ncbi_status"]
+            else:
+                as_name = canon_name
+                as_status = current["ncbi_status"]
+                canon_name = _format_name(name_pair)
+
+            txt = [
+                _indent(len(stack)),
+                canon_name,
+                " (",
+                as_status,
+                " from: ",
+                as_name,
+                ")",
+                tags.br(),
+            ]
+        else:
+            txt = [_indent(len(stack)), canon_name, tags.br()]
+
+        html += txt
+
+        # Is this taxon a parent for other taxa - if so add that taxon to the top of
+        # the stack, otherwise start looking for a next taxon to push onto the stack.
+        # If there is none at the top, pop and look down.
+        parent_id = current["ncbi_taxon_id"]
+        if parent_id in grouped:
+            stack.append(
+                ({"current": grouped[parent_id][0]}, {"next": grouped[parent_id][1:]})
+            )
+        else:
+            while stack:
+                push = stack.pop()
+                if push[1]["next"]:
+                    stack.append(
+                        ({"current": push[1]["next"][0]}, {"next": push[1]["next"][1:]})
+                    )
+                    break
+
+    return html
 
 
 def taxa_strip(name: str, rank: str) -> tuple[str, bool]:
